requirements.txt
requests
beautifulsoup4
pandas
numpy
scikit-learn
statsmodels
pyyaml
fuzzywuzzy
python-Levenshtein
matplotlib
seaborn
pytest